Version 1-12-83
pvConnection.cc: fix issue related to calc pvs

xygraph.cc: fix pv connection logic issue

button.cc button.h: change internal value to an int so bits may range from 0 to 31

+++ baselib/button.cc
@@ -280,7 +280,8 @@ activeButtonClass *bto = (activeButtonClass *) userarg;
 int st, sev;
 
   bto->controlValid = 1;
-  bto->curControlV = (short) pv->get_int();
+  //bto->curControlV = (short) pv->get_int();
+  bto->curControlV = pv->get_int();
 
   if ( bto->controlIsBit ) {
     bto->controlBit = ( ( bto->curControlV & ( 1 << bto->controlBitPos ) ) > 0 );
@@ -346,7 +347,8 @@ activeButtonClass *bto = (activeButtonClass *) userarg;
 int st, sev;
 
   bto->readValid = 1;
-  bto->curReadV = (short) pv->get_int();
+  //bto->curReadV = (short) pv->get_int();
+  bto->curReadV = pv->get_int();
 
   if ( bto->readIsBit ) {
     bto->readBit = ( ( bto->curReadV & ( 1 << bto->readBitPos ) ) > 0 );
@@ -2425,7 +2427,8 @@ void activeButtonClass::btnUp (
   int buttonNumber )
 {
 
-short value;
+//short value;
+int value;
 int stat;
 unsigned int uival;
 
@@ -2468,7 +2471,8 @@ void activeButtonClass::btnDown (
   int buttonNumber )
 {
 
-short value;
+//short value;
+int value;
 int stat, curControlBit;
 unsigned int uival;
 
@@ -2671,7 +2675,8 @@ void activeButtonClass::executeDeferred ( void ) {
 int ncc, nci, ncr, nrc, nri, nrr, ne, nd, nvc, nvi, nvu, ncolc, ncoli, ncolu;
 int stat, index, invisColor;
 
-short rv, cv;
+//short rv, cv;
+int rv, cv;
 char msg[79+1];
 
   if ( actWin->isIconified ) return;
@@ -2718,7 +2723,8 @@ char msg[79+1];
       return;
     }
 
-    cv = curControlV = (short) controlPvId->get_int();
+    //cv = curControlV = (short) controlPvId->get_int();
+    cv = curControlV = controlPvId->get_int();
 
     if ( controlIsBit ) {
       prevControlBit = controlBit = ( ( cv & ( 1 << controlBitPos ) ) > 0 );
@@ -2783,7 +2789,8 @@ char msg[79+1];
       return;
     }
 
-    rv = curReadV = (short) readPvId->get_int();
+    //rv = curReadV = (short) readPvId->get_int();
+    rv = curReadV = readPvId->get_int();
 
     if ( readIsBit ) {
       prevReadBit = readBit = ( ( rv & ( 1 << readBitPos ) ) > 0 );
@@ -3003,7 +3010,8 @@ int activeButtonClass::setProperty (
 
   if ( strcmp( prop, "controlValue" ) == 0 ) {
 
-    curControlV = (short) *value;
+    //curControlV = (short) *value;
+    curControlV = *value;
     needCtlRefresh = 1;
     actWin->appCtx->proc->lock();
     actWin->addDefExeNode( aglPtr );
@@ -3013,7 +3021,8 @@ int activeButtonClass::setProperty (
   }
   else if ( strcmp( prop, "readValue" ) == 0 ) {
 
-    curReadV = (short) *value;
+    //curReadV = (short) *value;
+    curReadV = *value;
     needReadRefresh = 1;
     actWin->appCtx->proc->lock();
     actWin->addDefExeNode( aglPtr );
